+++ packages/utils/src/misc.ts
@@ -6,11 +6,8 @@ import { isString } from './is';
  *
  * @param request The module path to resolve
  */
-export function bundlerSafeRequire(request: string): any {
-  // We need to do this check for global.module || module here because if we are calling this inside a
-  // active domain, global.module is undefined.
-  // tslint:disable-next-line
-  return ((getGlobalObject() as any).module || module).require(request);
+export function dynamicRequire(mod: NodeModule, request: string): any {
+  return mod.require(request);
 }
 
 /**