+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- [utils]: `bundlerSafeRequire` renamed to `dynamicRequire` now takes two arguments, first is should be `module`, second `request` / `moduleName`.
+
 ## 4.2.2
 
 - [core]: Several internal fixes regarding integration, exports and domain.